Best practices: only one class per file
Put the helper classes for the `Requirements_Test` each in a separate file in a new `Doubles` directory.
